moving required files to events folder
