+++ app/electron/events/launcher.js
